Hook local player camera rotation
