+++ src/launch/java/baritone/launch/mixins/MixinClientPlayerEntity.java
@@ -22,15 +22,18 @@ import baritone.api.IBaritone;
 import baritone.api.event.events.PlayerUpdateEvent;
 import baritone.api.event.events.SprintStateEvent;
 import baritone.api.event.events.type.EventState;
+import baritone.api.utils.Rotation;
 import baritone.behavior.LookBehavior;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.world.entity.player.Abilities;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 /**
  * @author Brady
@@ -51,6 +54,30 @@ public class MixinClientPlayerEntity {
         IBaritone baritone = BaritoneAPI.getProvider().getBaritoneForPlayer((LocalPlayer) (Object) this);
         if (baritone != null) {
             baritone.getGameEventHandler().onPlayerUpdate(new PlayerUpdateEvent(EventState.PRE));
+        }
+    }
+
+    @Inject(
+            method = "getViewYRot",
+            at = @At("HEAD"),
+            cancellable = true
+    )
+    private void getViewYRot(float partialTicks, CallbackInfoReturnable<Float> cir) {
+        Rotation visualRotation = this.getVisualRotation(partialTicks);
+        if (visualRotation != null) {
+            cir.setReturnValue(visualRotation.getYaw());
+        }
+    }
+
+    @Inject(
+            method = "getViewXRot",
+            at = @At("HEAD"),
+            cancellable = true
+    )
+    private void getViewXRot(float partialTicks, CallbackInfoReturnable<Float> cir) {
+        Rotation visualRotation = this.getVisualRotation(partialTicks);
+        if (visualRotation != null) {
+            cir.setReturnValue(visualRotation.getPitch());
         }
     }
 
@@ -119,5 +146,14 @@ public class MixinClientPlayerEntity {
             return false;
         }
         return instance.tryToStartFallFlying();
+    }
+
+    @Unique
+    private Rotation getVisualRotation(float partialTicks) {
+        IBaritone baritone = BaritoneAPI.getProvider().getBaritoneForPlayer((LocalPlayer) (Object) this);
+        if (baritone == null) {
+            return null;
+        }
+        return ((LookBehavior) baritone.getLookBehavior()).getVisualRotation(partialTicks);
     }
 }
