+++ src/launch/java/baritone/launch/mixins/MixinEntity.java
@@ -18,10 +18,7 @@
 package baritone.launch.mixins;
 
 import baritone.api.BaritoneAPI;
-import baritone.api.IBaritone;
 import baritone.api.event.events.RotationMoveEvent;
-import baritone.api.utils.Rotation;
-import baritone.behavior.LookBehavior;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.world.entity.Entity;
 import org.spongepowered.asm.mixin.Mixin;
@@ -30,7 +27,6 @@ import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(Entity.class)
 public class MixinEntity {
@@ -69,41 +65,5 @@ public class MixinEntity {
             this.xRot = this.motionUpdateRotationEvent.getOriginal().getPitch();
             this.motionUpdateRotationEvent = null;
         }
-    }
-
-    @Inject(
-            method = "getViewYRot",
-            at = @At("HEAD"),
-            cancellable = true
-    )
-    private void getViewYRot(float partialTicks, CallbackInfoReturnable<Float> cir) {
-        Rotation visualRotation = this.getVisualRotation(partialTicks);
-        if (visualRotation != null) {
-            cir.setReturnValue(visualRotation.getYaw());
-        }
-    }
-
-    @Inject(
-            method = "getViewXRot",
-            at = @At("HEAD"),
-            cancellable = true
-    )
-    private void getViewXRot(float partialTicks, CallbackInfoReturnable<Float> cir) {
-        Rotation visualRotation = this.getVisualRotation(partialTicks);
-        if (visualRotation != null) {
-            cir.setReturnValue(visualRotation.getPitch());
-        }
-    }
-
-    @Unique
-    private Rotation getVisualRotation(float partialTicks) {
-        if (!LocalPlayer.class.isInstance(this)) {
-            return null;
-        }
-        IBaritone baritone = BaritoneAPI.getProvider().getBaritoneForPlayer((LocalPlayer) (Object) this);
-        if (baritone == null) {
-            return null;
-        }
-        return ((LookBehavior) baritone.getLookBehavior()).getVisualRotation(partialTicks);
     }
 }
